extract bundle context to a module
