+++ src/snowflake/cli/plugins/nativeapp/bundle_context.py
@@ -1,0 +1,43 @@
+# Copyright (c) 2024 Snowflake Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+from pathlib import Path
+from typing import (
+    List,
+)
+
+from snowflake.cli.api.project.schemas.native_app.path_mapping import PathMapping
+from snowflake.cli.plugins.nativeapp.project_model import NativeAppProjectModel
+
+
+@dataclass
+class BundleContext:
+    package_name: str
+    artifacts: List[PathMapping]
+    project_root: Path
+    bundle_root: Path
+    deploy_root: Path
+    generated_root: Path
+
+    @staticmethod
+    def from_na_project(na: NativeAppProjectModel):
+        return BundleContext(
+            na.package_name,
+            na.artifacts,
+            na.project_root,
+            na.bundle_root,
+            na.deploy_root,
+            na.generated_root,
+        )
